WIP: Adding Missouri and updating puppet code.
First real PR so feedback welcome.

The code in mo_state is just a copy from Florida, so ignore it other than the tableau_click_selector_list setting up the list of arguments to be sent to puppet.py to get data from Tableau.

The Missouri dashboard was in ARCGIS https://mophep.maps.arcgis.com/apps/MapSeries/index.html?appid=8e01a5d8d8bd4b4f85add006f9e14a9d
   but has been moved to Tableau https://showmestrong.mo.gov/data-download/

The Excel data that they have is not as detailed as what is available in TableauNeedsClick
https://results.mo.gov/t/COVID19/views/COVID-19DataforDownload/MetricsbyTestDatebyCounty?%3AisGuestRedirectFromVizportal=y&%3Aembed=y

The page is protected by F5 to stop bots from accessing the site. Take a look in the cookies and the notes in can_tools/scrapers/puppet.py for more details.

Also, this Tableau sheet requires more clicks to get the data via the Download button. Updated the code to allow the state code to specify the click stream to follow to get to the data via tableau_click_selector_list.

One more fun challenge. When downloading data from Tableau's Data download option, it opens another tab that shows the first 1000 rows of data. There is a link on that new tab that allows you to download ALL the data. Clicking that opens another tab that will download the data. I was unable to find away to set the Page.setDownloadBehavior for this new page, hence, _special_tableau_handler code. This / Any special function can be added to the tableau_click_selector_list and that code will be run as part of the click process. In this case, it will get the download link from the new tab and download that code manually via the Requests module.

Lastly, Windows doesn't like the puppeteer / chromium options --single-process, --no-sandbox. I wrapped them in a os.name != 'nt' check to only include them in non-Windows environments.