Revert "Revert cdc test positivity change" (#237)
We've reevaluated the CDC dashboard and it seems that while they're not using percent_positive_7_day for their chart, they do use (the second to last data point from) it for the current test positivity value. We are going to reach out to CDC to see if we can get clarity on the two columns, but for now we're feeling more comfortable moving forward with the percent_positive_7_day column.