Add Prefect flow to generate one Parquet file per location
This is part of a series of explorations for improving the CAN data
pipeline.

In `services/prefect/flows/update_api_view.py`, the `UpdateParquetFiles`
flow generates one Parquet file containing all rows in the `covid_us`
view from PostgreSQL. This file is very large (~430 MB) and is
inconvenient for independent processing of data bounded to particular
locations.

We add a new flow `UpdateLocationParquetFiles` that instead generates
one Parquet file for each location. Resulting files range in size from
20 KB to 268 KB.

The flow queries the PostgreSQL database for a list of locations upon
each run, so editing the set of locations doesn't require code changes.