Apply generous timeouts to all prefect tasks
Apply generous timeouts to all prefect tasks to prevent zombie processes.

The prefect vm is crashing frequently. We believe one of the root causes is resource leaks from incorrectly handled processes that hang. As a first pass, we are adding generous timeouts for all tasks and removing any stale scrapers that are no longer needed.

These hour timeouts could be tightened in the future based on measure usage.