Remove State-Specific Vaccine Scrapers
Remove State-Specific Vaccine Scrapers

After an audit of state-specific vaccine scrapers we've come to the conclusion that the vast majority are no longer being updated. The few that remain are not providing useful new data.

We are removing all the state scrapers to shrink our maintenance surface and improve out prefect up time.